+++ jsonschema/golangci.next.jsonschema.json
@@ -250,7 +250,7 @@
             "gocyclo",
             "godot",
             "godox",
-            "goerr113",
+            "err113",
             "gofmt",
             "gofumpt",
             "goheader",
