+++ pkg/printers/testdata/in-report-data.json
@@ -135,7 +135,7 @@
         "Name": "godox"
       },
       {
-        "Name": "goerr113"
+        "Name": "err113"
       },
       {
         "Name": "gofmt",
